Guess low and result DOM manipulation
Guess low now works.
I changed it so that the result is displayed on the index page using DOM manipulation.
I have added a function that is called when the end-user clicks on "Guess 1-54".
The function creates a random number and uses an if statement checking if the number is between 1-54 if so it console logs a win for now, if its a number higher than 54 such as 55-100 it will declare a loss.

I had problems creating a function numberGenerator() that simply generates a number and changes the result DOM, I will try to implement it later on to help with redundant code.