Changed "betValue"
Instead of having an input I've put spans that when clicked they trigger an onclick function that changes the "betValue" based on which span they click.

I've used a global scope, defining the variable before they are called on the function so that the value of "betValue" can later be used to play the game and play bets.