Game functions
I have made it so that if you select a bet amount when you win or lose the overall money amount changes, and this manipulates the money shown to the user.
I need to make it so that the user can see the amount of money he is betting on the game.