Added background to bets
Added a new class called "betSelected" which changes the background to gray for now.

When you click on a bet value it adds the class to the selected bet and deletes the class from the other bet amounts, therefore making it so only the selected bet amount has a background difference.

I tried to use "classList.toggle" but found myself unsuccessful linking them properly because if the user clicks on the value more than once it could still have the "betValue" 10 but the background would toggle off for the span selected.