Slot Machine Concept
Added a new slot machine concept, allows the user to buy a slot machine in the game by pressing the button and it will slowly generate them money using a setInterval.

The code needs to be improved and on the index I need to add a button image and also description for the player so that they know what it does.