Added slot machine counter
I've made sure the user can't buy a slot machine if they cannot afford one. I've also added a counter to count the amount of machines the user has bought.
I append a child to the newly created element and a function is called which adds a display for a new slot machine.